refactor: implement state management with Cubit and di
