+++ lib/presentation/screens/account/cubits/account_state.dart
@@ -62,6 +62,7 @@ class AccountLoaded extends AccountState {
     notificationsEnabled,
     selectedLanguage,
     appVersion,
+    activeBottomSheet,
   ];
 }
