+++ lib/presentation/screens/account/delete_account/delete_account_bottom_sheet.dart
@@ -1,8 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sellio_mobile/core/design_system/constants/app_strings.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
 import 'package:sellio_mobile/core/design_system/widgets/buttons/button.dart';
 import 'package:sellio_mobile/core/design_system/widgets/sellio_bottom_sheet.dart';
+import '../cubits/BottomSheetType.dart';
+import '../cubits/account_cubit.dart';
+import '../cubits/account_state.dart';
 
 class DeleteAccountBottomSheet extends StatelessWidget {
   final VoidCallback onDeleteAccount;
@@ -16,13 +20,33 @@ class DeleteAccountBottomSheet extends StatelessWidget {
     return SellioBottomSheet.show(
       context: context,
       isScrollControlled: true,
-      child: DeleteAccountBottomSheet(onDeleteAccount: onDeleteAccount),
+        child: BlocProvider.value(
+          value: context.read<AccountCubit>(),
+          child: DeleteAccountBottomSheet(onDeleteAccount: onDeleteAccount),
+        ),
     );
   }
 
   @override
   Widget build(BuildContext context) {
-    return Column(
+    return BlocListener<AccountCubit, AccountState>(
+        listenWhen: (previous, current) {
+          if (current is AccountLoaded &&
+              current.activeBottomSheet == BottomSheetType.none) {
+            if (previous is AccountLoaded) {
+              return previous.activeBottomSheet != BottomSheetType.none;
+            }
+            if (previous is AccountActionLoading) {
+              return previous.previousState.activeBottomSheet !=
+                  BottomSheetType.none;
+            }
+          }
+          return false;
+        },
+        listener: (context, state) {
+          Navigator.of(context).pop();
+        },
+        child: Column(
           mainAxisSize: MainAxisSize.min,
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
@@ -46,6 +70,6 @@ class DeleteAccountBottomSheet extends StatelessWidget {
               fullWidth: true,
             ),
           ],
-        );
+        ));
   }
 }
