+++ lib/presentation/screens/account/language/change_language_bottom_sheet.dart
@@ -1,8 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sellio_mobile/core/design_system/constants/app_strings.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
-import 'package:sellio_mobile/core/design_system/widgets/sellio_bottom_sheet.dart';
 import '../../../../core/design_system/widgets/buttons/button.dart';
+import '../../../../core/design_system/widgets/sellio_bottom_sheet.dart';
+import '../cubits/BottomSheetType.dart';
+import '../cubits/account_cubit.dart';
+import '../cubits/account_state.dart';
 
 class ChangeLanguageBottomSheet extends StatefulWidget {
   final String selectedLanguage;
@@ -25,9 +29,12 @@ class ChangeLanguageBottomSheet extends StatefulWidget {
     return SellioBottomSheet.show(
       context: context,
       isScrollControlled: true,
-      child: ChangeLanguageBottomSheet(
-        onSave: onSave,
-        selectedLanguage: selectedLanguage,
+      child: BlocProvider.value(
+        value: context.read<AccountCubit>(),
+        child: ChangeLanguageBottomSheet(
+          onSave: onSave,
+          selectedLanguage: selectedLanguage,
+        ),
       ),
     );
   }
@@ -40,64 +47,86 @@ class _ChangeLanguageBottomSheetState extends State<ChangeLanguageBottomSheet> {
   @override
   void initState() {
     super.initState();
-    currentSelectedLanguage = widget.selectedLanguage;
+    final accountState = context.read<AccountCubit>().state;
+    if (accountState is AccountLoaded) {
+      currentSelectedLanguage = accountState.selectedLanguage;
+    } else {
+      currentSelectedLanguage = AppStrings.english;
+    }
     _updateSaveButtonState();
   }
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          AppStrings.changeLanguage,
-          style: context.theme.typography.textTheme.titleMedium,
-        ),
-        const SizedBox(height: 24),
-        RadioGroup<String>(
-          groupValue: currentSelectedLanguage,
-          onChanged: (String? value) {
-            if (value != null) {
-              setState(() {
-                currentSelectedLanguage = value;
-              });
-              _updateSaveButtonState();
+    return BlocListener<AccountCubit, AccountState>(
+        listenWhen: (previous, current) {
+          if (current is AccountLoaded &&
+              current.activeBottomSheet == BottomSheetType.none) {
+            if (previous is AccountLoaded) {
+              return previous.activeBottomSheet != BottomSheetType.none;
             }
-          },
-          child: Row(
-            children: [
-              Expanded(
-                child: _buildLanguageOption(
-                  language: AppStrings.english,
-                  context: context,
-                ),
+            if (previous is AccountActionLoading) {
+              return previous.previousState.activeBottomSheet !=
+                  BottomSheetType.none;
+            }
+          }
+          return false;
+        },
+        listener: (context, state) {
+          Navigator.of(context).pop();
+        },
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(
+              AppStrings.changeLanguage,
+              style: context.theme.typography.textTheme.titleMedium,
+            ),
+            const SizedBox(height: 24),
+            RadioGroup<String>(
+              groupValue: currentSelectedLanguage,
+              onChanged: (String? value) {
+                if (value != null) {
+                  setState(() {
+                    currentSelectedLanguage = value;
+                  });
+                  _updateSaveButtonState();
+                }
+              },
+              child: Row(
+                children: [
+                  Expanded(
+                    child: _buildLanguageOption(
+                      language: AppStrings.english,
+                      context: context,
+                    ),
+                  ),
+                  const SizedBox(width: 8),
+                  Expanded(
+                    child: _buildLanguageOption(
+                      language: AppStrings.arabic,
+                      context: context,
+                    ),
+                  ),
+                ],
               ),
-              const SizedBox(width: 8),
-              Expanded(
-                child: _buildLanguageOption(
-                  language: AppStrings.arabic,
-                  context: context,
-                ),
+            ),
+            const SizedBox(height: 24),
+            SizedBox(
+              width: double.infinity,
+              child: SellioButton(
+                text: AppStrings.save,
+                onTap: _isFormValid
+                    ? () {
+                        widget.onSave(currentSelectedLanguage);
+                      }
+                    : null,
+                isEnabled: _isFormValid,
               ),
-            ],
-          ),
-        ),
-        const SizedBox(height: 24),
-        SizedBox(
-          width: double.infinity,
-          child: SellioButton(
-            text: AppStrings.save,
-            onTap: _isFormValid
-                ? () {
-                    widget.onSave(currentSelectedLanguage);
-                  }
-                : null,
-            isEnabled: _isFormValid,
-          ),
-        ),
-      ],
-    );
+            ),
+          ],
+        ));
   }
 
   void _updateSaveButtonState() {
