+++ lib/presentation/screens/main/main_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:sellio_mobile/presentation/screens/account/account_screen.dart';
 import '../../../core/design_system/widgets/bottom_nav_bar.dart';
-import '../account_screen.dart';
 import '../cart_screen.dart';
 import '../home/home_screen.dart';
 import '../customize_your_product_screen/CustomizeYourProductScreen.dart';
